requirements.txt
Flask
Flask-Cors
rembg
Pillow
onnxruntime
gunicorn # প্রোডাকশন সার্ভার হিসেবে Gunicorn ব্যবহার করা হবে